Rearranging full repair macro
