+++ src/module/macros.ts
@@ -43,10 +43,11 @@ export { renderMacroTemplate, renderMacroHTML } from "./macros/render"
 export { prepareActivationMacro } from "./macros/activation"
 export { prepareItemMacro } from "./macros/item"
 export { prepareOverchargeMacro } from "./macros/overcharge"
-export { stabilizeMacro } from "./macros/stabilize"
 export { prepareStatMacro } from "./macros/stat"
 export { prepareTechMacro } from "./macros/tech"
 export { prepareTextMacro } from "./macros/text"
+export { stabilizeMacro } from "./macros/stabilize"
+export { fullRepairMacro } from "./macros/full-repair"
 
 const lp = LANCER.log_prefix;
 
@@ -980,42 +981,6 @@ export async function prepareStructureMacro(
 
   // Hand it off to the actor to structure
   await actor.structure(reroll_data);
-}
-
-export function fullRepairMacro(a: string) {
-  // Determine which Actor to speak as
-  let actor = getMacroSpeaker(a);
-  if (!actor) return Promise.reject();
-
-  return new Promise<boolean>((resolve, reject) => {
-    new Dialog({
-      title: `FULL REPAIR - ${actor?.name}`,
-      content: `<h3>Are you sure you want to fully repair the ${actor?.data.type} ${actor?.name}?`,
-      buttons: {
-        submit: {
-          icon: '<i class="fas fa-check"></i>',
-          label: "Yes",
-          callback: async _dlg => {
-            // Gotta typeguard the actor again
-            if (!actor) {
-              return reject();
-            }
-
-            await actor.full_repair();
-            prepareTextMacro(a, "REPAIRED", `Notice: ${actor.name} has been fully repaired.`);
-            resolve(true);
-          },
-        },
-        cancel: {
-          icon: '<i class="fas fa-times"></i>',
-          label: "No",
-          callback: async () => resolve(false),
-        },
-      },
-      default: "submit",
-      close: () => resolve(false),
-    }).render(true);
-  });
 }
 
 /**
